+++ modules/vm-monitoring/outputs.tf
@@ -1,14 +1,14 @@
 output "data_collection_rule_id" {
-  description = "Id of the Azure Monitor Data Collection Rule"
-  value       = local.data_collection_rule_id
+  description = "ID of the Azure Monitor Data Collection Rule"
+  value       = azurerm_monitor_data_collection_rule.dcr.id
 }
 
 output "data_collection_rule_name" {
   description = "Name of the Azure Monitor Data Collection Rule"
-  value       = local.name
+  value       = azurerm_monitor_data_collection_rule.dcr.name
 }
 
-output "data_collection_rule_data" {
-  description = "JSON data of the Azure Monitor Data Collection Rule"
-  value       = local.data_collection_config
+output "data_collection_rule" {
+  description = "Azure Monitor Data Collection Rule object"
+  value       = azurerm_monitor_data_collection_rule.dcr
 }