+++ modules/vm-monitoring/versions.tf
@@ -3,15 +3,11 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.0"
+      version = "~> 3.15"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"
       version = "~> 1.1"
-    }
-    null = {
-      source  = "hashicorp/null"
-      version = ">= 3"
     }
   }
 }